+++ tests/manager/test_rest_api.py
@@ -2,11 +2,15 @@ import json
 
 import numpy
 import pytest
-from django.shortcuts import reverse
+from rest_framework.reverse import reverse
 
 from topobank.manager.models import Surface, Topography
 from topobank.testing.factories import SurfaceFactory, TagFactory
-from topobank.testing.utils import assert_dict_equal, assert_dicts_equal
+from topobank.testing.utils import (
+    ASSERT_EQUAL_IGNORE_VALUE,
+    assert_dict_equal,
+    assert_dicts_equal,
+)
 from topobank.users.anonymous import get_anonymous_user
 
 
@@ -36,6 +40,9 @@ def test_surface_retrieve_routes(
     response = api_client.get(reverse("manager:surface-api-list"))
     assert response.status_code == 400
 
+    topography_api_list_url = reverse(
+        "manager:topography-api-list", request=response.wsgi_request
+    )
     surface1_dict = {
         "category": None,
         "creator": f"http://testserver/users/api/user/{user.id}/",
@@ -47,6 +54,7 @@ def test_surface_retrieve_routes(
         "creation_datetime": surface1.creation_datetime.astimezone().isoformat(),
         "modification_datetime": surface1.modification_datetime.astimezone().isoformat(),
         "attachments": surface1.attachments.get_absolute_url(response.wsgi_request),
+        "topographies": f"{topography_api_list_url}?surface={surface1.id}",
     }
     if hasattr(Surface, "publication"):
         surface1_dict["publication"] = None
@@ -66,15 +74,14 @@ def test_surface_retrieve_routes(
                 "other_users": [],
             },
             "attachments": topo1.attachments.get_absolute_url(response.wsgi_request),
-            "deepzoom": None,
-            "bandwidth_lower": None,
-            "bandwidth_upper": None,
+            "bandwidth_lower": topo1.bandwidth_lower,
+            "bandwidth_upper": topo1.bandwidth_upper,
             "creator": f"http://testserver/users/api/user/{user.id}/",
-            "datafile_format": None,
+            "datafile_format": topo1.datafile_format,
             "description": "description1",
             "detrend_mode": "height",
             "fill_undefined_data_mode": "do-not-fill",
-            "has_undefined_data": None,
+            "has_undefined_data": False,
             "height_scale": 0.296382712790741,
             "height_scale_editable": False,
             "instrument_name": "",
@@ -87,21 +94,23 @@ def test_surface_retrieve_routes(
             "resolution_x": 256,
             "resolution_y": 256,
             "short_reliability_cutoff": None,
-            "size_editable": True,
-            "size_x": 10.0,
-            "size_y": 10.0,
+            "size_editable": False,
+            "size_x": 10000.0,
+            "size_y": 10000.0,
             "surface": f"http://testserver/manager/api/surface/{surface1.id}/",
-            "unit": "µm",
+            "unit": "nm",
             "unit_editable": False,
             "url": f"http://testserver/manager/api/topography/{topo1.id}/",
             "duration": None,
             "error": None,
             "id": topo1.id,
-            "squeezed_datafile": None,
+            "datafile": ASSERT_EQUAL_IGNORE_VALUE,
+            "squeezed_datafile": ASSERT_EQUAL_IGNORE_VALUE,
+            "deepzoom": ASSERT_EQUAL_IGNORE_VALUE,
             "tags": [],
             "task_progress": 0.0,
-            "task_state": "no",
-            "thumbnail": None,
+            "task_state": "pe",
+            "thumbnail": ASSERT_EQUAL_IGNORE_VALUE,
             "channel_names": [],
             "data_source": 0,
             "is_metadata_complete": True,
@@ -120,6 +129,7 @@ def test_surface_retrieve_routes(
         "creation_datetime": surface2.creation_datetime.astimezone().isoformat(),
         "modification_datetime": surface2.modification_datetime.astimezone().isoformat(),
         "attachments": surface2.attachments.get_absolute_url(response.wsgi_request),
+        "topographies": f"{topography_api_list_url}?surface={surface2.id}",
     }
     if hasattr(Surface, "publication"):
         surface2_dict["publication"] = None
@@ -139,15 +149,14 @@ def test_surface_retrieve_routes(
                 "other_users": [],
             },
             "attachments": topo2.attachments.get_absolute_url(response.wsgi_request),
-            "deepzoom": None,
-            "bandwidth_lower": None,
-            "bandwidth_upper": None,
+            "bandwidth_lower": topo2.bandwidth_lower,
+            "bandwidth_upper": topo2.bandwidth_upper,
             "creator": f"http://testserver/users/api/user/{user.id}/",
-            "datafile_format": None,
+            "datafile_format": topo2.datafile_format,
             "description": "description2",
             "detrend_mode": "height",
             "fill_undefined_data_mode": "do-not-fill",
-            "has_undefined_data": None,
+            "has_undefined_data": False,
             "height_scale": 2.91818e-08,
             "height_scale_editable": False,
             "instrument_name": "",
@@ -157,24 +166,26 @@ def test_surface_retrieve_routes(
             "is_periodic_editable": True,
             "measurement_date": "2018-01-02",
             "name": "Example 4 - Default",
-            "resolution_x": 305,
-            "resolution_y": 75,
+            "resolution_x": 75,
+            "resolution_y": 305,
             "short_reliability_cutoff": None,
             "size_editable": False,
-            "size_x": 112.80791,
-            "size_y": 27.73965,
+            "size_y": 112.80791e-6,
+            "size_x": 27.73965e-6,
             "surface": f"http://testserver/manager/api/surface/{surface2.id}/",
-            "unit": "µm",
+            "unit": "m",
             "unit_editable": False,
             "url": f"http://testserver/manager/api/topography/{topo2.id}/",
             "duration": None,
             "error": None,
             "id": topo2.id,
-            "squeezed_datafile": None,
+            "datafile": ASSERT_EQUAL_IGNORE_VALUE,
+            "squeezed_datafile": ASSERT_EQUAL_IGNORE_VALUE,
+            "deepzoom": ASSERT_EQUAL_IGNORE_VALUE,
+            "thumbnail": ASSERT_EQUAL_IGNORE_VALUE,
             "tags": [],
             "task_progress": 0.0,
-            "task_state": "no",
-            "thumbnail": None,
+            "task_state": "pe",
             "channel_names": [],
             "data_source": 0,
             "is_metadata_complete": True,
@@ -194,17 +205,12 @@ def test_surface_retrieve_routes(
     if is_authenticated:
         assert response.status_code == 200
         data = json.loads(json.dumps(response.data))  # Convert OrderedDict to dict
-        if "topography_set" in data:
-            for t in data["topography_set"]:
-                del t["datafile"]  # datafile has an S3 hash which is difficult to mock
         assert_dict_equal(data, surface1_dict)
 
         response = api_client.get(
             f"{reverse('manager:topography-api-list')}?surface={surface1.id}&permissions=yes&attachments=yes"
         )
         data = response.data
-        for t in data:
-            del t["datafile"]  # datafile has an S3 hash which is difficult to mock
         assert_dicts_equal(data, surface1_topographies_dict)
     else:
         # Anonymous user does not have access by default
@@ -217,17 +223,12 @@ def test_surface_retrieve_routes(
     if is_authenticated:
         assert response.status_code == 200
         data = json.loads(json.dumps(response.data))  # Convert OrderedDict to dict
-        if "topography_set" in data:
-            for t in data["topography_set"]:
-                del t["datafile"]  # datafile has an S3 hash which is difficult to mock
         assert_dict_equal(data, surface2_dict)
 
         response = api_client.get(
             f"{reverse('manager:topography-api-list')}?surface={surface2.id}&permissions=yes&attachments=yes"
         )
         data = response.data
-        for t in data:
-            del t["datafile"]  # datafile has an S3 hash which is difficult to mock
         assert_dicts_equal(data, surface2_topographies_dict)
     else:
         # Anonymous user does not have access by default
@@ -258,14 +259,14 @@ def test_topography_retrieve_routes(
     )
 
     topo1_dict = {
-        "bandwidth_lower": None,
-        "bandwidth_upper": None,
+        "bandwidth_lower": topo1.bandwidth_lower,
+        "bandwidth_upper": topo1.bandwidth_upper,
         "creator": f"http://testserver/users/api/user/{user.id}/",
-        "datafile_format": None,
+        "datafile_format": topo1.datafile_format,
         "description": "description1",
         "detrend_mode": "height",
         "fill_undefined_data_mode": "do-not-fill",
-        "has_undefined_data": None,
+        "has_undefined_data": False,
         "height_scale": 0.296382712790741,
         "height_scale_editable": False,
         "instrument_name": "",
@@ -278,38 +279,44 @@ def test_topography_retrieve_routes(
         "resolution_x": 256,
         "resolution_y": 256,
         "short_reliability_cutoff": None,
-        "size_editable": True,
-        "size_x": 10.0,
-        "size_y": 10.0,
+        "size_editable": False,
+        "size_x": 10000.0,
+        "size_y": 10000.0,
         "surface": f"http://testserver/manager/api/surface/{topo1.surface.id}/",
-        "unit": "µm",
+        "unit": "nm",
         "unit_editable": False,
         "url": f"http://testserver/manager/api/topography/{topo1.id}/",
         "tags": [],
         "task_progress": 0.0,
         "task_state": "pe",
-        "thumbnail": None,
-        "squeezed_datafile": None,
+        "thumbnail": ASSERT_EQUAL_IGNORE_VALUE,
         "is_metadata_complete": True,
         "id": topo1.id,
         "error": None,
         "duration": None,
         "data_source": 0,
-        "channel_names": [],
+        "channel_names": [
+            ["ZSensor", "nm"],
+            ["AmplitudeError", None],
+            ["Phase", None],
+            ["Height", "nm"],
+        ],
         "creation_datetime": topo1.creation_datetime.astimezone().isoformat(),
         "modification_datetime": topo1.modification_datetime.astimezone().isoformat(),
         "attachments": topo1.attachments.get_absolute_url(response.wsgi_request),
-        "deepzoom": None,
+        "deepzoom": ASSERT_EQUAL_IGNORE_VALUE,
+        "datafile": ASSERT_EQUAL_IGNORE_VALUE,
+        "squeezed_datafile": ASSERT_EQUAL_IGNORE_VALUE,
     }
     topo2_dict = {
-        "bandwidth_lower": None,
-        "bandwidth_upper": None,
+        "bandwidth_lower": topo2.bandwidth_lower,
+        "bandwidth_upper": topo2.bandwidth_upper,
         "creator": f"http://testserver/users/api/user/{user.id}/",
-        "datafile_format": None,
+        "datafile_format": topo2.datafile_format,
         "description": "description2",
         "detrend_mode": "height",
         "fill_undefined_data_mode": "do-not-fill",
-        "has_undefined_data": None,
+        "has_undefined_data": False,
         "height_scale": 2.91818e-08,
         "height_scale_editable": False,
         "instrument_name": "",
@@ -319,21 +326,20 @@ def test_topography_retrieve_routes(
         "is_periodic_editable": True,
         "measurement_date": "2018-01-02",
         "name": "Example 4 - Default",
-        "resolution_x": 305,
-        "resolution_y": 75,
+        "resolution_x": 75,
+        "resolution_y": 305,
         "short_reliability_cutoff": None,
         "size_editable": False,
-        "size_x": 112.80791,
-        "size_y": 27.73965,
+        "size_y": 112.80791e-6,
+        "size_x": 27.73965e-6,
         "surface": f"http://testserver/manager/api/surface/{topo2.surface.id}/",
-        "unit": "µm",
+        "unit": "m",
         "unit_editable": False,
         "url": f"http://testserver/manager/api/topography/{topo2.id}/",
         "tags": [],
         "task_progress": 0.0,
         "task_state": "pe",
-        "thumbnail": None,
-        "squeezed_datafile": None,
+        "thumbnail": ASSERT_EQUAL_IGNORE_VALUE,
         "is_metadata_complete": True,
         "id": topo2.id,
         "error": None,
@@ -343,19 +349,20 @@ def test_topography_retrieve_routes(
         "creation_datetime": topo2.creation_datetime.astimezone().isoformat(),
         "modification_datetime": topo2.modification_datetime.astimezone().isoformat(),
         "attachments": topo2.attachments.get_absolute_url(response.wsgi_request),
-        "deepzoom": None,
+        "deepzoom": ASSERT_EQUAL_IGNORE_VALUE,
+        "datafile": ASSERT_EQUAL_IGNORE_VALUE,
+        "squeezed_datafile": ASSERT_EQUAL_IGNORE_VALUE,
     }
 
     if is_authenticated:
         assert response.status_code == 200
         data = json.loads(json.dumps(response.data))  # Convert OrderedDict to dict
-        del data["datafile"]  # datafile has an S3 hash which is difficult to mock
         # topo1 is updated but the get command, because it is triggering file inspection
         topo1 = Topography.objects.get(pk=topo1.id)
         topo1_dict["modification_datetime"] = (
             topo1.modification_datetime.astimezone().isoformat()
         )
-        assert data == topo1_dict
+        assert_dict_equal(data, topo1_dict)
     else:
         # Anonymous user does not have access by default
         assert response.status_code == 404
@@ -366,13 +373,12 @@ def test_topography_retrieve_routes(
     if is_authenticated:
         assert response.status_code == 200
         data = json.loads(json.dumps(response.data))  # Convert OrderedDict to dict
-        del data["datafile"]  # datafile has an S3 hash which is difficult to mock
         # topo2 is updated but the get command, because it is triggering file inspection
         topo2 = Topography.objects.get(pk=topo2.id)
         topo2_dict["modification_datetime"] = (
             topo2.modification_datetime.astimezone().isoformat()
         )
-        assert data == topo2_dict
+        assert_dict_equal(data, topo2_dict)
     else:
         # Anonymous user does not have access by default
         assert response.status_code == 404
@@ -404,7 +410,7 @@ def test_create_surface_routes(api_client, two_users, handle_usage_statistics):
     assert s.creator.name == user1.name
 
 
-@pytest.mark.django_db(transaction=True)
+@pytest.mark.django_db
 def test_delete_surface_routes(api_client, two_users, handle_usage_statistics):
     (user1, user2), (surface1, surface2, surface3) = two_users
     topo1, topo2, topo3 = Topography.objects.all()
@@ -643,7 +649,7 @@ def test_tag_retrieve_routes(api_client, two_users, handle_usage_statistics):
 
     # Anonymous user should not be able to see the tag
     response = api_client.get(
-        reverse("manager:tag-api-detail", kwargs=dict(name=st.path))
+        reverse("manager:tag-api-detail", kwargs=dict(name=st.name))
     )
     assert response.status_code == 403
 
@@ -651,7 +657,7 @@ def test_tag_retrieve_routes(api_client, two_users, handle_usage_statistics):
     # surfaces that are tagged, hence the tag does not exist for her
     api_client.force_login(user1)
     response = api_client.get(
-        f"{reverse('manager:tag-api-detail', kwargs=dict(name=st.path))}?surfaces=yes"
+        f"{reverse('manager:tag-api-detail', kwargs=dict(name=st.name))}?surfaces=yes"
     )
     assert response.status_code == 403
 
@@ -673,11 +679,14 @@ def test_tag_retrieve_routes(api_client, two_users, handle_usage_statistics):
     # User 2 has access to all surfaces inside the tag
     api_client.force_login(user2)
     response = api_client.get(
-        reverse("manager:tag-api-detail", kwargs=dict(name=st.path))
+        reverse("manager:tag-api-detail", kwargs=dict(name=st.name))
     )
     assert response.data["name"] == st.name
 
     response = api_client.get(f"{reverse('manager:surface-api-list')}?tag={st.name}")
+    topography_api_list_url = reverse(
+        "manager:topography-api-list", request=response.wsgi_request
+    )
     assert_dicts_equal(
         response.data,
         [
@@ -694,6 +703,7 @@ def test_tag_retrieve_routes(api_client, two_users, handle_usage_statistics):
                 "attachments": surface2.attachments.get_absolute_url(
                     response.wsgi_request
                 ),
+                "topographies": f"{topography_api_list_url}?surface={surface2.id}",
             },
             {
                 "url": surface3.get_absolute_url(response.wsgi_request),
@@ -708,6 +718,7 @@ def test_tag_retrieve_routes(api_client, two_users, handle_usage_statistics):
                 "attachments": surface3.attachments.get_absolute_url(
                     response.wsgi_request
                 ),
+                "topographies": f"{topography_api_list_url}?surface={surface3.id}",
             },
         ],
     )
@@ -833,3 +844,45 @@ def test_create_topography_with_blank_name_fails(
         },
     )
     assert response.status_code == 400
+
+
+def test_set_permissions(api_client, user_alice, user_bob, handle_usage_statistics):
+    surface = SurfaceFactory(creator=user_alice)
+    surface.grant_permission(user_alice, "full")
+
+    api_client.force_login(user_bob)
+    response = api_client.get(
+        reverse("manager:surface-api-detail", kwargs=dict(pk=surface.id))
+    )
+    assert response.status_code == 404
+
+    api_client.force_login(user_alice)
+    response = api_client.patch(
+        reverse("manager:set-permissions", kwargs=dict(pk=surface.id)),
+        [{"user": dict(id=user_bob.id), "permission": "full"}],
+    )
+    assert response.status_code == 204
+
+    api_client.force_login(user_bob)
+    response = api_client.get(
+        reverse("manager:surface-api-detail", kwargs=dict(pk=surface.id))
+    )
+    assert response.status_code == 200
+
+    response = api_client.patch(
+        reverse("manager:set-permissions", kwargs=dict(pk=surface.id)),
+        [{"user": dict(id=user_bob.id), "permission": "no-access"}],
+    )
+    assert response.status_code == 405  # Cannot remove permission from logged in user
+
+    response = api_client.patch(
+        reverse("manager:set-permissions", kwargs=dict(pk=surface.id)),
+        [{"user": dict(id=user_alice.id), "permission": "no-access"}],
+    )
+    assert response.status_code == 204  # Cannot remove permission from logged in user
+
+    api_client.force_login(user_alice)
+    response = api_client.get(
+        reverse("manager:surface-api-detail", kwargs=dict(pk=surface.id))
+    )
+    assert response.status_code == 404