+++ topobank/manager/models.py
@@ -9,8 +9,8 @@ import os.path
 import sys
 import tempfile
 from collections import defaultdict
+from typing import List
 
-import dateutil.parser
 import django.dispatch
 import matplotlib.pyplot
 import numpy as np
@@ -22,9 +22,11 @@ from django.core.files import File
 from django.core.files.base import ContentFile
 from django.core.validators import MinValueValidator
 from django.db import models
+from django.db.models import Q
 from rest_framework.reverse import reverse
 from SurfaceTopography.Container.SurfaceContainer import SurfaceContainer
 from SurfaceTopography.Exceptions import UndefinedDataError
+from SurfaceTopography.Metadata import InstrumentParametersModel
 from SurfaceTopography.Support.UnitConversion import get_unit_conversion_factor
 
 from ..authorization.mixins import PermissionMixin
@@ -132,11 +134,6 @@ class SubjectMixin:
 class Tag(tm.TagTreeModel, SubjectMixin):
     """This is the common tag model for surfaces and topographies."""
 
-    class TagMeta:
-        force_lowercase = True
-        # not needed yet
-        # autocomplete_view = 'manager:autocomplete-tags'
-
     _user = None
 
     def authorize_user(
@@ -173,7 +170,7 @@ class Tag(tm.TagTreeModel, SubjectMixin):
     def is_shared(self, user: settings.AUTH_USER_MODEL) -> bool:
         return True  # Tags are generally shared, but the surfaces may not
 
-    def get_authorized_user(self):
+    def get_authorized_user(self) -> settings.AUTH_USER_MODEL:
         return self._user
 
     def get_related_surfaces(self):
@@ -186,7 +183,7 @@ class Tag(tm.TagTreeModel, SubjectMixin):
             )
         return Surface.objects.for_user(self._user).filter(tags=self.id)
 
-    def get_children(self):
+    def get_children(self) -> List[str]:
         def make_child(tag_name):
             tag_suffix = tag_name[len(self.name) + 1 :]
             name, rest = (tag_suffix + "/").split("/", maxsplit=1)
@@ -218,8 +215,8 @@ class Tag(tm.TagTreeModel, SubjectMixin):
                 "to restrict user permissions."
             )
         return Surface.objects.for_user(self._user).filter(
-            tags__path__startswith=self.path
-        )
+            Q(tags=self) | Q(tags__name__startswith=f"{self.name}/")
+        ).distinct()
 
     def get_properties(self, kind=None):
         """
@@ -756,9 +753,9 @@ class Topography(PermissionMixin, TaskStateModel, SubjectMixin):
             ]
 
             # `instrument_parameters` is special as it can contain non-significant entries
-            if self._clean_instrument_parameters(
-                self.instrument_parameters
-            ) != self._clean_instrument_parameters(old_obj.instrument_parameters):
+            if InstrumentParametersModel(
+                **self.instrument_parameters
+            ) != InstrumentParametersModel(**old_obj.instrument_parameters):
                 changed_fields += ["instrument_parameters"]
 
             # We need to refresh if any of the significant fields changed during this save
@@ -847,54 +844,16 @@ class Topography(PermissionMixin, TaskStateModel, SubjectMixin):
         """
         return self.permissions.get_for_user(user) is not None
 
-    @staticmethod
-    def _clean_instrument_parameters(params):
-        cleaned_params = {}
-        if params is None:
-            return cleaned_params
-
-        def _clean_value_unit_pair(r):
-            cleaned_r = None
-            if "value" in r and "unit" in r:
-                # Value/unit pair is complete
-                try:
-                    cleaned_r = {"value": float(r["value"]), "unit": r["unit"]}
-                except KeyError:
-                    # 'value' or 'unit' does not exist - should not happen
-                    pass
-                except TypeError:
-                    # Value is None
-                    pass
-                except ValueError:
-                    # Value cannot be converted to float
-                    pass
-            return cleaned_r
-
-        # Check completeness of resolution parameters
-        for key in ["resolution", "tip_radius"]:
-            try:
-                r = _clean_value_unit_pair(params[key])
-            except KeyError:
-                pass
-            else:
-                if r is not None:
-                    cleaned_params[key] = r
-
-        return cleaned_params
-
     @property
-    def _instrument_info(self):
-        # We need to idiot-check the parameters JSON so surface topography does not complain
-        # Would it be better to use JSON Schema for this? Or should we simply have dedicated database fields?
-        params = self._clean_instrument_parameters(self.instrument_parameters)
-
+    def instrument_info(self):
         # Build dictionary with instrument information from database... this may override data provided by the
         # topography reader
         return {
             "instrument": {
                 "name": self.instrument_name,
-                "type": self.instrument_type,
-                "parameters": params,
+                "parameters": InstrumentParametersModel(
+                    **self.instrument_parameters
+                ).model_dump(exclude_none=True),
             }
         }
 
@@ -932,7 +891,7 @@ class Topography(PermissionMixin, TaskStateModel, SubjectMixin):
             reader_kwargs["unit"] = self.unit
 
         # Populate instrument information
-        reader_kwargs["info"] = self._instrument_info
+        reader_kwargs["info"] = self.instrument_info
 
         # Eventually get topography from module "SurfaceTopography" using the given keywords
         topo = reader.topography(**reader_kwargs)
@@ -992,7 +951,7 @@ class Topography(PermissionMixin, TaskStateModel, SubjectMixin):
             toporeader = get_topography_reader(
                 self.squeezed_datafile.file, format=SQUEEZED_DATAFILE_FORMAT
             )
-            topo = toporeader.topography(info=self._instrument_info)
+            topo = toporeader.topography(info=self.instrument_info)
             # In the squeezed format, these things are already applied/included:
             # unit, scaling, detrending, physical sizes
             # so don't need to provide them to the .topography() method
@@ -1447,9 +1406,7 @@ class Topography(PermissionMixin, TaskStateModel, SubjectMixin):
         if populate_initial_metadata:
             # Measurement time
             try:
-                self.measurement_date = dateutil.parser.parse(
-                    channel.info["acquisition_time"]
-                )
+                self.measurement_date = channel.info["acquisition_time"]
             except:  # noqa: E722
                 pass